+++ LearningProgress/LearningProgressInterfaces/BindingModels/DisciplineLessonTaskStudentAcceptBindingModels.cs
@@ -32,7 +32,7 @@ namespace LearningProgressInterfaces.BindingModels
         [Required(ErrorMessage = "required")]
         public DateTime DateAccept { get; set; }
 
-        public decimal Score { get; set; }
+        public double Score { get; set; }
 
         public string Comment { get; set; }
